requirements.txt
requests
selenium==4.14.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
validators==0.22.0
python-dotenv==1.0.0